chore: Workaround focus handling differences
