+++ src/Uno.UI/UI/Xaml/Controls/Frame/Frame.legacy.cs
@@ -212,9 +212,7 @@ public partial class Frame : ContentControl
 			CurrentEntry.Instance = page;
 		}
 
-		MoveFocusFromCurrentContent();
-
-		Content = CurrentEntry.Instance;
+		SetContent(CurrentEntry.Instance);
 
 		if (IsNavigationStackEnabled)
 		{
@@ -313,53 +311,6 @@ public partial class Frame : ContentControl
 					entry.SourcePageType,
 					null
 				));
-	}
-
-	/// <summary>
-	/// In case the current page contains a focused element,
-	/// we need to move the focus out of the page.
-	/// </summary>
-	/// <remarks>
-	/// In UWP this is done automatically as the elements are unloaded,
-	/// but due to the control lifecycle differences in Uno the focus move multiple times
-	/// as controls are unloaded in "layers" and it could also not move outside this Frame,
-	/// as the Parent would already be unassigned during the OnUnloaded execution.
-	/// </remarks>
-	private void MoveFocusFromCurrentContent()
-	{
-		if (Content is not UIElement uiElement)
-		{
-			return;
-		}
-		uiElement.IsLeavingFrame = true;
-		try
-		{
-			var focusManager = VisualTree.GetFocusManagerForElement(this);
-			if (focusManager?.FocusedElement is not { } focusedElement)
-			{
-				return;
-			}
-
-			var parent = VisualTreeHelper.GetParent(focusedElement);
-			while (parent is not null && parent != this)
-			{
-				parent = VisualTreeHelper.GetParent(parent);
-			}
-
-			var inCurrentPage = parent == this;
-
-			if (inCurrentPage)
-			{
-				// Set the focus on the next focusable element.
-				focusManager.SetFocusOnNextFocusableElement(FocusState.Programmatic, true);
-
-				(focusedElement as Control)?.UpdateFocusState(FocusState.Unfocused);
-			}
-		}
-		finally
-		{
-			uiElement.IsLeavingFrame = false;
-		}
 	}
 
 	private void OnSourcePageTypeChangedLegacy(DependencyPropertyChangedEventArgs e)
